Added GML3solidMockData
