+++ modules/extension/xsd/xsd-gml3/src/test/java/org/geotools/gml3/bindings/GML3MockData.java
@@ -32,6 +32,7 @@ import org.geotools.referencing.CRS;
 import org.geotools.xml.XSD;
 import org.opengis.feature.simple.SimpleFeature;
 import org.opengis.feature.simple.SimpleFeatureType;
+import org.opengis.geometry.primitive.Solid;
 import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -640,134 +641,6 @@ public class GML3MockData {
         
         return surface;
     }
-    
-    /* 
-     * for SolidTypeBindingTest
-     * */
-    public static Element compositeSurfaceWithValidity(Document document, Node parent) {
-        Element compositeSurface = element(qName("CompositeSurface"), document, parent);
-        Element surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* right surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("2 0 0 2 2 0 2 2 2 2 0 2 2 0 0"));
-        }
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* back surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("2 2 0 0 2 0 0 2 2 2 2 2 2 2 0"));
-        }
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* left surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("0 0 0 0 0 2 0 2 2 0 2 0 0 0 0"));
-        }
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* front surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("0 0 0 2 0 0 2 0 2 0 0 2 0 0 0"));
-        }
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* lower surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("0 0 0 0 2 0 2 2 0 2 0 0 0 0 0"));
-        }
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        {
-            /* upper surface */
-            Element polygon = element(qName("Polygon"), document, surfaceMember);
-            polygon.setAttribute("srsDimension", "3");
-    
-            Element exterior = element(qName("exterior"), document, polygon);
-            Element linearRing = element(qName("LinearRing"), document, exterior);
-            linearRing.setAttribute("srsDimension", "3");
-            Element posList = element(qName("posList"), document, linearRing);
-            linearRing.appendChild(posList);
-            posList.appendChild(document.createTextNode("0 0 2 2 0 2 2 2 2 0 2 2 0 0 2"));
-        }
-        
-        
-        return compositeSurface;
-    }
-    
-    public static Element compositeSurface(Document document, Node parent) {
-        Element compositeSurface = element(qName("CompositeSurface"), document, parent);
-        Element surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        polygonWithPosList3D(document, surfaceMember, true);
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        polygonWithPosList3D(document, surfaceMember, true);
-        
-        surfaceMember = element(qName("surfaceMember"), document, compositeSurface);
-        polygonWithPosList3D(document, surfaceMember, true);
-        
-        return compositeSurface;
-    }
-    
-    public static Element solid(Document document, Node parent) {
-        return solid(document, parent, false);
-    }
-    
-    public static Element solid(Document document, Node parent, boolean withInterior) {
-        Element solid = element(qName("Solid"), document, parent);
-        solid.setAttribute("srsDimension", "3");
-        
-        Element exterior = element(qName("exterior"), document, solid);
-        //compositeSurface(document, exterior);
-        compositeSurfaceWithValidity(document, exterior);
-        
-        if(withInterior) {
-            Element interior = element(qName("interior"), document, solid);
-            compositeSurface(document, interior);
-        }
-        
-        return solid;
-    }
-    /* */
     
     public static Element feature(Document document, Node parent) {
         Element feature = element(TEST.TestFeature, document, parent);
